% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MurineMatchTx-package}
\alias{MurineMatchTx}
\alias{MurineMatchTx-package}
\title{MurineMatchTx package}
\description{
Internal package level configuration for shared imports and check notes.
}
\author{
\strong{Maintainer}: Dian Wang \email{wangdian020803@gmail.com}

Authors:
\itemize{
  \item Guiwen Guan \email{guanguiwen@bjmu.edu.cn}
}

}
\keyword{internal}
