% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DEG_heatmap.R
\name{plot_DEG_heatmap}
\alias{plot_DEG_heatmap}
\title{Heatmap for DEG Analysis Based on TCGA High and Low Expression Groups}
\usage{
plot_DEG_heatmap(
  deg_data,
  mouse_files,
  col_names = NULL,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  color_palette = c(`1` = "#ff7676", `-1` = "#66d4ff", `NA` = "white"),
  width = 7.9,
  height = 5.95,
  output_path = "./DEG/"
)
}
\arguments{
\item{deg_data}{List. Output 'processed data' from 'prepare_DEG_heatmap_data()', containing processed TCGA and mouse DEG data.}

\item{mouse_files}{Named list. The original input 'mouse_files' from 'merge_DEG_datasets()', needed for column names.}

\item{col_names}{Character vector. Custom column names for the heatmap. Default uses 'names(mouse_files)'.}

\item{cluster_rows}{Logical. Whether to cluster rows (default: FALSE).}

\item{cluster_cols}{Logical. Whether to cluster columns (default: FALSE).}

\item{color_palette}{Named vector. Colors for the heatmap (default: c("1" = "#ff7676", "-1" = "#66d4ff", "NA" = "white")).}

\item{width}{Integer. The width of pdf (default: 7.9).}

\item{height}{Integer. The height of pdf (default: 5.95).}

\item{output_path}{Character. File path for saving the plot (default: "./DEG/").}
}
\value{
A matrix ('merge_edger') containing merged DEG expression data as the input object of 'plot_DEG_barplot()'
}
\description{
This function generates a heatmap of log fold changes (logFC)
for differentially expressed genes (DEGs) across multiple datasets.
}
\examples{
\dontrun{
# example usage
merge_edger <- plot_DEG_heatmap(deg_data = processed_data, mouse_files = mouse_files)
}

}
