% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA_tinyarray.R
\name{plot_PCA_tinyarray}
\alias{plot_PCA_tinyarray}
\title{PCA Plot Using 'tinyarray::draw_pca()'}
\usage{
plot_PCA_tinyarray(pca_data, batch_correction = TRUE, colors = NULL)
}
\arguments{
\item{pca_data}{List. Output from 'prepare_PCA_data()', containing merged expression data.}

\item{batch_correction}{Logical. Whether to use batch-corrected data (default: 'TRUE').}

\item{colors}{Character vector. Colors for different groups.}
}
\description{
This function performs PCA analysis and visualizes the results using 'tinyarray::draw_pca()'.
}
\examples{
\dontrun{
# draw PCA
plot_PCA_tinyarray(pca_data = pca_data, batch_correction = T)
}

}
