% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TME_boxplot.R
\name{plot_TME_boxplot}
\alias{plot_TME_boxplot}
\title{Plot Tumor Microenvironment (TME) Boxplots for Multiple Signatures}
\usage{
plot_TME_boxplot(
  eset_tpm,
  group_as,
  condition_list,
  width = 10,
  height = 6,
  output_path = "./TME_boxplots"
)
}
\arguments{
\item{eset_tpm}{Data frame. TPM-transformed RNA-seq expression matrix, output from 'plot_TME_barplot()'.}

\item{group_as}{Character vector. 'Identical' with the 'group_as' in 'plot_TME_barplot()'.}

\item{condition_list}{Named list. Sample count for each mouse model (e.g., 'list("DEN" = 6, "GSExxx" = 8)').}

\item{width}{Integer. Width of saved plots (default: 10).}

\item{height}{Integer. Height of saved plots (default: 6).}

\item{output_path}{Character. Directory path to save boxplots (default: "./TME_boxplots").}
}
\value{
Saves multiple boxplots and returns a list of ggplot objects.
}
\description{
This function computes immune infiltration scores using CIBERSORT and generates
boxplots for multiple immune signatures.
}
\examples{
\dontrun{
group_as <- c(rep("HBV_normal", 3), rep("HBV_tumor", 3), rep("DEN_normal", 4), rep("DEN_tumor", 4))
condition_list <- list("HBV" = 6, "DEN" = 8)
plot_TME_boxplot(eset_tpm = TME_barplot_result$eset_tpm, group_as = group_as, condition_list = condition_list)
}

}
