% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TROMscore_boxplot.R
\name{plot_TROMscore_boxplot}
\alias{plot_TROMscore_boxplot}
\title{Plot TROM Score Boxplot and Tileplot}
\usage{
plot_TROMscore_boxplot(
  merge_long,
  output_path = "./similarity/TROM_boxplot",
  combine_plots = TRUE
)
}
\arguments{
\item{merge_long}{Data frame. Output from 'plot_TROMscore_heatmap()', containing 'TROM scores'.}

\item{output_path}{Character. File path to save the figure (default: "./similarity/TROM_boxplot").}

\item{combine_plots}{Logical. Whether to combine tileplot & boxplot into one figure (default: 'TRUE').}
}
\value{
Saves the plot(s) and returns the ggplot object.
}
\description{
This function generates two subplots:
\enumerate{
\item Boxplot of transcriptional similarity scores ('TROM score').
\item Tileplot showing the mean similarity scores per condition.
}
}
\examples{
\dontrun{
plot_TROMscore_boxplot(
  merge_long = heatmap_results$merge_long,
  output_path = "./similarity/TROM_boxplot",
  combine_plots = TRUE
)
}

}
