% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TROMscore_heatmap.R
\name{plot_TROMscore_heatmap}
\alias{plot_TROMscore_heatmap}
\title{Plot TROM Score Heatmap for TCGA and Mouse Models}
\usage{
plot_TROMscore_heatmap(
  dm_trom_del_backgroud_clinic,
  trom_matrix,
  tcga_sample_size,
  tcga_normal_count,
  model_group_count,
  condition,
  clinical_vars,
  cnv_genes,
  group_all,
  width = 11.84,
  height = 7.11,
  output_path = "./similarity/TROM_heatmap.pdf",
  cluster_rows = FALSE,
  cluster_cols = FALSE
)
}
\arguments{
\item{dm_trom_del_backgroud_clinic}{Data frame. Output from 'compute_similarity_matrix()', containing 'TROM scores'.}

\item{trom_matrix}{Data frame. 'Raw ws.trom similarity matrix' before background removal.Output from 'compute_similarity_matrix()'.}

\item{tcga_sample_size}{Integer. Total number of TCGA samples.}

\item{tcga_normal_count}{Integer. Number of TCGA normal samples.}

\item{model_group_count}{Integer. Total number of Mouse model samples.}

\item{condition}{Named list. User-defined sample counts per mouse model condition (e.g., 'list("HBV pten p53 ko" = 5, "DEN Cl4" = 21)').}

\item{clinical_vars}{Character vector. Clinical variables used in 'prepare_dm_trom_del_backgroud_clinic()'.}

\item{cnv_genes}{Character vector. Copy number variation (CNV) genes used in 'prepare_dm_trom_del_backgroud_clinic()'.}

\item{group_all}{The group information of TCGA and all mouse models from 'prepare_PCA_data()'.}

\item{width}{Integer. The width of PDF (default: 11.84).}

\item{height}{Integer. The height of PDF (default: 7.11).}

\item{output_path}{Character. File path for saving the heatmap (default: "./similarity/TROM_heatmap.pdf").}

\item{cluster_rows}{Logical. Whether to cluster rows (default: 'FALSE').}

\item{cluster_cols}{Logical. Whether to cluster columns (default: 'FALSE').}
}
\value{
A list containing:
\describe{
\item{heatmap_obj}{The generated heatmap.}
\item{merge_long}{Processed similarity scores for downstream analysis.}
}
}
\description{
This function visualizes transcriptional similarity scores between TCGA samples and mouse models.
}
\details{
\itemize{
\item Returns 'heatmap_obj' (heatmap plot) and 'merge_long' (processed similarity scores).
\item 'merge_long' is used as the input for 'plot_TROMscore_boxplot()'.
}
}
\examples{
\dontrun{
heatmap_results <- plot_TROMscore_heatmap(
  dm_trom_del_backgroud_clinic = similarity_results$dm_trom_del_backgroud_clinic,
  trom_matrix = similarity_results$similarity_matrix,
  tcga_sample_size = 424,
  tcga_normal_count = 50,
  model_group_count = 52,
  condition = list("HBV pten p53 ko" = 5, "DEN Cl4" = 16),
  clinical_vars = c("Hepatitis B", "Non-Alcoholic"),
  cnv_genes = c("TP53", "PTEN"),
  group_all = pca_data$group_all,
  output_path = "./similarity/TROM_heatmap.pdf"
)
}

}
